+++ resources/views/tasks/group_by_task_lists.blade.php
@@ -125,6 +125,8 @@
             </div>
         </div>
     </div>
-    <script src="{{ asset('assets/js/pages/group-by-task-lists.js') }}"></script>
+@endsection
 
+@section('page_scripts')
+    <script src="{{ asset('assets/js/pages/group-by-task-lists.js') }}"></script>
 @endsection